revert moving files to subdirectory
