+++ src/lib/holocene/accordion.story.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import type { Hst as HST } from '@histoire/plugin-svelte';
   import Accordion from './accordion.svelte';
-  import Button from '../button.svelte';
-  import TableHeaderRow from '../table/table-header-row.svelte';
-  import TableRow from '../table/table-row.svelte';
-  import Table from '../table/table.svelte';
+  import Button from './button.svelte';
+  import TableHeaderRow from './table/table-header-row.svelte';
+  import TableRow from './table/table-row.svelte';
+  import Table from './table/table.svelte';
 
   export let Hst: HST;
 </script>
