--- src/lib/holocene/accordion/index.ts
@@ -1,3 +1,0 @@
-import Accordion from './accordion.svelte';
-
-export default Accordion;
